Fixed shadow warnings in all files
#823 Multiple variable shadowing warnings in private/bitset_new.h

+++ test/test_bitset_new_default_element_type.cpp
@@ -1143,7 +1143,7 @@ namespace
     //*************************************************************************
     TEST(test_to_ulong_min)
     {
-      ETL_CONSTEXPR14 etl::bitset<etl::integral_limits<unsigned long>::bits> data((unsigned long)etl::integral_limits<unsigned long>::min);
+      ETL_CONSTEXPR14 etl::bitset<etl::integral_limits<unsigned long>::bits> data(etl::integral_limits<unsigned long>::min);
       unsigned long value = data.to_ulong();
 
       CHECK_EQUAL(std::numeric_limits<unsigned long>::min(), value);
@@ -1152,7 +1152,7 @@ namespace
     //*************************************************************************
     TEST(test_to_ulong_max)
     {
-      ETL_CONSTEXPR14 etl::bitset<etl::integral_limits<unsigned long>::bits> data((unsigned long)etl::integral_limits<unsigned long>::max);
+      ETL_CONSTEXPR14 etl::bitset<etl::integral_limits<unsigned long>::bits> data(etl::integral_limits<unsigned long>::max);
       unsigned long value = data.to_ulong();
 
       CHECK_EQUAL(std::numeric_limits<unsigned long>::max(), value);
@@ -1161,14 +1161,14 @@ namespace
     //*************************************************************************
     TEST(test_to_ulong_overflow)
     {
-      ETL_CONSTEXPR14 etl::bitset<etl::integral_limits<unsigned long>::bits + 1> data((unsigned long)etl::integral_limits<unsigned long>::min);
+      ETL_CONSTEXPR14 etl::bitset<etl::integral_limits<unsigned long>::bits + 1> data(etl::integral_limits<unsigned long>::min);
       CHECK_THROW(data.to_ulong(), etl::bitset_overflow);
     }
 
     //*************************************************************************
     TEST(test_to_ullong_min)
     {
-      ETL_CONSTEXPR14 etl::bitset<etl::integral_limits<unsigned long long>::bits> data((unsigned long long)etl::integral_limits<unsigned long long>::min);
+      ETL_CONSTEXPR14 etl::bitset<etl::integral_limits<unsigned long long>::bits> data(etl::integral_limits<unsigned long long>::min);
       unsigned long long value = data.to_ullong();
 
       CHECK_EQUAL(std::numeric_limits<unsigned long>::min(), value);
@@ -1177,7 +1177,7 @@ namespace
     //*************************************************************************
     TEST(test_to_ullong_max)
     {
-      ETL_CONSTEXPR14 etl::bitset<etl::integral_limits<unsigned long long>::bits> data((unsigned long long)etl::integral_limits<unsigned long long>::max);
+      ETL_CONSTEXPR14 etl::bitset<etl::integral_limits<unsigned long long>::bits> data(etl::integral_limits<unsigned long long>::max);
       unsigned long long value = data.to_ullong();
 
       CHECK_EQUAL(std::numeric_limits<unsigned long long>::max(), value);
@@ -1186,7 +1186,7 @@ namespace
     //*************************************************************************
     TEST(test_to_ullong_overflow)
     {
-      ETL_CONSTEXPR14 etl::bitset<etl::integral_limits<unsigned long long>::bits + 1> data((unsigned long long)etl::integral_limits<unsigned long long>::min);
+      ETL_CONSTEXPR14 etl::bitset<etl::integral_limits<unsigned long long>::bits + 1> data(etl::integral_limits<unsigned long long>::min);
       CHECK_THROW(data.to_ullong(), etl::bitset_overflow);
     }
